Added option to prevent backup when current size is smaller

+++ Classes/General.cs
@@ -650,13 +650,16 @@ namespace JDP {
             return bytes.Length / 2;
         }
 
-        public static void BackupThreadList() {
+        public static void BackupThreadList(bool checkSize = false) {
             try {
                 string path = Path.Combine(Settings.GetSettingsDirectory(), Settings.ThreadsFileName);
                 if (!File.Exists(path)) return;
-                string[] lines = File.ReadAllLines(path);
-                if (lines.Length < 1) return;
-                File.WriteAllLines(path + ".bak", lines);
+                var backupInfo = new FileInfo(path + ".bak");
+                if (!checkSize || !backupInfo.Exists || new FileInfo(path).Length >= backupInfo.Length) {
+                    string[] lines = File.ReadAllLines(path);
+                    if (lines.Length < 1) return;
+                    File.WriteAllLines(path + ".bak", lines);
+                }
             }
             catch (Exception ex) {
                 Logger.Log(ex.ToString());
